+++ crmeb/src/main/java/com/zbkj/crmeb/store/response/StoreProductCartProductInfoResponse.java
@@ -109,7 +109,7 @@ public class StoreProductCartProductInfoResponse implements Serializable {
 //    private Boolean merUse;
 
     @ApiModelProperty(value = "获得积分")
-    private BigDecimal giveIntegral;
+    private Integer giveIntegral;
 
     @ApiModelProperty(value = "成本价")
     private BigDecimal cost;
